Rename klaus.ctags to klaus.ctagsutils
Avoid name clash with python-ctags
