+++ bin/klaus
@@ -66,7 +66,7 @@ def main():
         args.site_name = '%s:%d' % (args.host, args.port)
 
     if args.ctags != 'none':
-        from klaus.ctags import check_have_exuberant_ctags
+        from klaus.ctagsutils import check_have_exuberant_ctags
         if not check_have_exuberant_ctags():
             print >> sys.stderr, "ERROR: Exuberant ctags not installed (or 'ctags' binary isn't *Exuberant* ctags)"
             return 1
